Update README.md
We have created a decentralized voting system using smart contracts in solitude language in this we have defined a contract named voting in this code. In the contract, there is a struct named Voter, which contains three fields - weight, voted, and vote. A voter's weight is stored in the weight field, their vote is stored in the vote field, and their vote is stored in the vote field.